backend/requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# API Documentation
drf-spectacular==0.27.0

# Async Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Storage
boto3==1.34.14
django-storages==1.14.2

# Security
cryptography==41.0.7
python-dotenv==1.0.0

# QR Code
qrcode==7.4.2
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2
django-extensions==3.2.3

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Export
openpyxl==3.1.2

# PDF Generation
reportlab==4.0.8

# Development
ipython>=7.0.0
